Ignore pixels with 0.0 value, don't calibrate to energies or add to histograms.
Former-commit-id: d431148bef5c53baf525b1817624e77c1674dedb

+++ STFC/HexitecGigEProject/trunk/HxtProcessingLibTest/mainwindow.cpp
@@ -26,8 +26,8 @@ MainWindow::MainWindow(QWidget *parent) :
            processingBufferGenerator, SLOT(handleConfigureProcessing(bool,const char*,const char*,const char*)));
    connect(processingForm, SIGNAL(configureProcessing(int, int, uint16_t*, const char*, const char*, const char*)),
            processingBufferGenerator, SLOT(handleConfigureProcessing(int, int, uint16_t*, const char*, const char*, const char*)));
-   connect(processingForm, SIGNAL(configureProcessing(bool, long long, long long, long long, bool, const char *, const char *, const char *)),
-           processingBufferGenerator, SLOT(handleConfigureProcessing(bool, long long, long long, long long, bool, const char *, const char *, const char *)));
+   connect(processingForm, SIGNAL(configureProcessing(bool, long long, long long, double, bool, const char *, const char *, const char *)),
+           processingBufferGenerator, SLOT(handleConfigureProcessing(bool, long long, long long, double, bool, const char *, const char *, const char *)));
 
    connect(processingForm, SIGNAL(imageStarted(const char*, int)),
            processingBufferGenerator, SLOT(handleImageStarted(const char *, int)));
